Added an optional NoNegative parameter to doForwardEuler. This is untested.
deterministic.m is supposed to use doForwardEuler to simulate a four dimensional
system of ODEs that model phage lysis.

TODO: it is NOT working! Code compiles and runs, but
results from doForwardEuler are not what is expected
and disagrees with BOTH ode45 and sage's desolve_system_rk4()
solver for the same system. ode45 and desolve_system_rk4()
give the same equilibrium values, but the dynamics seem
different. NO CLUE.